Refactor Instance Methods Try 2 (#1254)
* Moved methods from instance_types to instance_controller and a helper file
