+++ pkg/controller/instance/instance_controller.go
@@ -17,11 +17,14 @@ package instance
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
 	"reflect"
 	"time"
+
+	"k8s.io/apimachinery/pkg/util/uuid"
 
 	"github.com/thoas/go-funk"
 	appsv1 "k8s.io/api/apps/v1"
@@ -46,6 +49,11 @@ import (
 	"github.com/kudobuilder/kudo/pkg/engine/task"
 	"github.com/kudobuilder/kudo/pkg/engine/workflow"
 	"github.com/kudobuilder/kudo/pkg/util/kudo"
+)
+
+const (
+	snapshotAnnotation           = "kudo.dev/last-applied-instance-state"
+	instanceCleanupFinalizerName = "kudo.dev.instance.cleanup"
 )
 
 // Reconciler reconciles an Instance object.
@@ -170,7 +178,7 @@ func (r *Reconciler) Reconcile(request ctrl.Request) (ctrl.Result, error) {
 
 	if !instance.IsDeleting() {
 		if _, hasCleanupPlan := ov.Spec.Plans[kudov1beta1.CleanupPlanName]; hasCleanupPlan {
-			if instance.TryAddFinalizer() {
+			if tryAddFinalizer(instance) {
 				log.Printf("InstanceController: Adding finalizer on instance %s/%s", instance.Namespace, instance.Name)
 			}
 		}
@@ -180,13 +188,13 @@ func (r *Reconciler) Reconcile(request ctrl.Request) (ctrl.Result, error) {
 
 	// ---------- 3. Check if we should start execution of new plan ----------
 
-	planToBeExecuted, err := instance.GetPlanToBeExecuted(ov)
+	planToBeExecuted, err := getPlanToBeExecuted(instance, ov)
 	if err != nil {
 		return reconcile.Result{}, err
 	}
 	if planToBeExecuted != nil {
 		log.Printf("InstanceController: Going to start execution of plan %s on instance %s/%s", kudo.StringValue(planToBeExecuted), instance.Namespace, instance.Name)
-		err = instance.StartPlanExecution(kudo.StringValue(planToBeExecuted), ov)
+		err = startPlanExecution(instance, kudo.StringValue(planToBeExecuted), ov)
 		if err != nil {
 			return reconcile.Result{}, r.handleError(err, instance, oldInstance)
 		}
@@ -264,7 +272,7 @@ func updateInstance(instance *kudov1beta1.Instance, oldInstance *kudov1beta1.Ins
 
 	// update instance metadata if finalizer is removed
 	// because Kubernetes might immediately delete the instance, this has to be the last instance update
-	if instance.TryRemoveFinalizer() {
+	if tryRemoveFinalizer(instance) {
 		log.Printf("InstanceController: Removing finalizer on instance %s/%s", instance.Namespace, instance.Name)
 		if err := client.Update(context.TODO(), instance); err != nil {
 			log.Printf("InstanceController: Error when removing instance finalizer. %v", err)
@@ -420,4 +428,301 @@ func pipesMap(planName string, plan *v1beta1.Plan, tasks []v1beta1.Task, emeta *
 	}
 
 	return pipes, nil
+}
+
+// startPlanExecution mark plan as to be executed
+// this modifies the instance.Status as well as instance.Metadata.Annotation (to save snapshot if needed)
+func startPlanExecution(i *v1beta1.Instance, planName string, ov *v1beta1.OperatorVersion) error {
+	if i.NoPlanEverExecuted() || isUpgradePlan(planName) {
+		ensurePlanStatusInitialized(i, ov)
+	}
+
+	// update status of the instance to reflect the newly starting plan
+	notFound := true
+	for planIndex, v := range i.Status.PlanStatus {
+		if v.Name == planName {
+			// update plan status
+			notFound = false
+			planStatus := i.Status.PlanStatus[planIndex]
+			planStatus.Set(v1beta1.ExecutionPending)
+			planStatus.UID = uuid.NewUUID()
+			for j, p := range v.Phases {
+				planStatus.Phases[j].Set(v1beta1.ExecutionPending)
+				for k := range p.Steps {
+					i.Status.PlanStatus[planIndex].Phases[j].Steps[k].Set(v1beta1.ExecutionPending)
+				}
+			}
+
+			i.Status.PlanStatus[planIndex] = planStatus // we cannot modify item in map, we need to reassign here
+
+			// update activePlan and instance status
+			i.Status.AggregatedStatus.Status = v1beta1.ExecutionPending
+			i.Status.AggregatedStatus.ActivePlanName = planName
+
+			break
+		}
+	}
+	if notFound {
+		return &v1beta1.InstanceError{Err: fmt.Errorf("asked to execute a plan %s but no such plan found in instance %s/%s", planName, i.Namespace, i.Name), EventName: kudo.String("PlanNotFound")}
+	}
+
+	err := saveSnapshot(i)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// ensurePlanStatusInitialized initializes plan status for all plans this instance supports
+// it does not trigger run of any plan
+// it either initializes everything for a fresh instance without any status or tries to adjust status after OV was updated
+func ensurePlanStatusInitialized(i *v1beta1.Instance, ov *v1beta1.OperatorVersion) {
+	if i.Status.PlanStatus == nil {
+		i.Status.PlanStatus = make(map[string]v1beta1.PlanStatus)
+	}
+
+	for planName, plan := range ov.Spec.Plans {
+		planStatus := &v1beta1.PlanStatus{
+			Name:   planName,
+			Status: v1beta1.ExecutionNeverRun,
+			Phases: make([]v1beta1.PhaseStatus, 0),
+		}
+
+		existingPlanStatus, planExists := i.Status.PlanStatus[planName]
+		if planExists {
+			planStatus.SetWithMessage(existingPlanStatus.Status, existingPlanStatus.Message)
+		}
+		for _, phase := range plan.Phases {
+			phaseStatus := &v1beta1.PhaseStatus{
+				Name:   phase.Name,
+				Status: v1beta1.ExecutionNeverRun,
+				Steps:  make([]v1beta1.StepStatus, 0),
+			}
+			existingPhaseStatus, phaseExists := v1beta1.PhaseStatus{}, false
+			if planExists {
+				for _, oldPhase := range existingPlanStatus.Phases {
+					if phase.Name == oldPhase.Name {
+						existingPhaseStatus = oldPhase
+						phaseExists = true
+						phaseStatus.SetWithMessage(existingPhaseStatus.Status, existingPhaseStatus.Message)
+					}
+				}
+			}
+			for _, step := range phase.Steps {
+				stepStatus := v1beta1.StepStatus{
+					Name:   step.Name,
+					Status: v1beta1.ExecutionNeverRun,
+				}
+				if phaseExists {
+					for _, oldStep := range existingPhaseStatus.Steps {
+						if step.Name == oldStep.Name {
+							stepStatus.SetWithMessage(oldStep.Status, oldStep.Message)
+						}
+					}
+				}
+				phaseStatus.Steps = append(phaseStatus.Steps, stepStatus)
+			}
+			planStatus.Phases = append(planStatus.Phases, *phaseStatus)
+		}
+		i.Status.PlanStatus[planName] = *planStatus
+	}
+}
+
+// isUpgradePlan returns true if this could be an upgrade plan - this is just an approximation because deploy plan can be used for both
+func isUpgradePlan(planName string) bool {
+	return planName == v1beta1.DeployPlanName || planName == v1beta1.UpgradePlanName
+}
+
+// getPlanToBeExecuted returns name of the plan that should be executed
+func getPlanToBeExecuted(i *v1beta1.Instance, ov *v1beta1.OperatorVersion) (*string, error) {
+	if i.IsDeleting() {
+		// we have a cleanup plan
+		plan := selectPlan([]string{v1beta1.CleanupPlanName}, ov)
+		if plan != nil {
+			if planStatus := i.PlanStatus(*plan); planStatus != nil {
+				if !planStatus.Status.IsRunning() {
+					if planStatus.Status.IsFinished() {
+						// we already finished the cleanup plan
+						return nil, nil
+					}
+					return plan, nil
+				}
+			}
+		}
+	}
+
+	if i.GetPlanInProgress() != nil { // we're already running some plan
+		return nil, nil
+	}
+
+	// new instance, need to run deploy plan
+	if i.NoPlanEverExecuted() {
+		return kudo.String(v1beta1.DeployPlanName), nil
+	}
+
+	// did the instance change so that we need to run deploy/upgrade/update plan?
+	instanceSnapshot, err := snapshotSpec(i)
+	if err != nil {
+		return nil, err
+	}
+	if instanceSnapshot == nil {
+		// we don't have snapshot -> we never run deploy, also we cannot run update/upgrade. This should never happen
+		return nil, &v1beta1.InstanceError{Err: fmt.Errorf("unexpected state: no plan is running, no snapshot present - this should never happen :) for instance %s/%s", i.Namespace, i.Name), EventName: kudo.String("UnexpectedState")}
+	}
+	if instanceSnapshot.OperatorVersion.Name != i.Spec.OperatorVersion.Name {
+		// this instance was upgraded to newer version
+		log.Printf("Instance: instance %s/%s was upgraded from %s to %s operatorVersion", i.Namespace, i.Name, instanceSnapshot.OperatorVersion.Name, i.Spec.OperatorVersion.Name)
+		plan := selectPlan([]string{v1beta1.UpgradePlanName, v1beta1.UpdatePlanName, v1beta1.DeployPlanName}, ov)
+		if plan == nil {
+			return nil, &v1beta1.InstanceError{Err: fmt.Errorf("supposed to execute plan because instance %s/%s was upgraded but none of the deploy, upgrade, update plans found in linked operatorVersion", i.Namespace, i.Name), EventName: kudo.String("PlanNotFound")}
+		}
+		return plan, nil
+	}
+	// did instance parameters change, so that the corresponding plan has to be triggered?
+	if !reflect.DeepEqual(instanceSnapshot.Parameters, i.Spec.Parameters) {
+		// instance updated
+		log.Printf("Instance: instance %s/%s has updated parameters from %v to %v", i.Namespace, i.Name, instanceSnapshot.Parameters, i.Spec.Parameters)
+		paramDiff := parameterDiff(instanceSnapshot.Parameters, i.Spec.Parameters)
+		paramDefinitions := getParamDefinitions(paramDiff, ov)
+		plan := planNameFromParameters(paramDefinitions, ov)
+		if plan == nil {
+			return nil, &v1beta1.InstanceError{Err: fmt.Errorf("supposed to execute plan because instance %s/%s was updated but none of the deploy, update plans found in linked operatorVersion", i.Namespace, i.Name), EventName: kudo.String("PlanNotFound")}
+		}
+		return plan, nil
+	}
+	return nil, nil
+}
+
+// planNameFromParameters determines what plan to run based on params that changed and the related trigger plans
+func planNameFromParameters(params []v1beta1.Parameter, ov *v1beta1.OperatorVersion) *string {
+	for _, p := range params {
+		// TODO: if the params have different trigger plans, we always select first here which might not be ideal
+		if p.Trigger != "" && selectPlan([]string{p.Trigger}, ov) != nil {
+			return kudo.String(p.Trigger)
+		}
+	}
+	return selectPlan([]string{v1beta1.UpdatePlanName, v1beta1.DeployPlanName}, ov)
+}
+
+// getParamDefinitions retrieves parameter metadata from OperatorVersion CRD
+func getParamDefinitions(params map[string]string, ov *v1beta1.OperatorVersion) []v1beta1.Parameter {
+	defs := []v1beta1.Parameter{}
+	for p1 := range params {
+		for _, p2 := range ov.Spec.Parameters {
+			if p2.Name == p1 {
+				defs = append(defs, p2)
+			}
+		}
+	}
+	return defs
+}
+
+// parameterDiff returns map containing all parameters that were removed or changed between old and new
+func parameterDiff(old, new map[string]string) map[string]string {
+	diff := make(map[string]string)
+
+	for key, val := range old {
+		// If a parameter was removed in the new spec
+		if _, ok := new[key]; !ok {
+			diff[key] = val
+		}
+	}
+
+	for key, val := range new {
+		// If new spec parameter was added or changed
+		if v, ok := old[key]; !ok || v != val {
+			diff[key] = val
+		}
+	}
+
+	return diff
+}
+
+// selectPlan returns nil if none of the plan exists, otherwise the first one in list that exists
+func selectPlan(possiblePlans []string, ov *v1beta1.OperatorVersion) *string {
+	for _, n := range possiblePlans {
+		if _, ok := ov.Spec.Plans[n]; ok {
+			return kudo.String(n)
+		}
+	}
+	return nil
+}
+
+// SaveSnapshot stores the current spec of Instance into the snapshot annotation
+// this information is used when executing update/upgrade plans, this overrides any snapshot that existed before
+func saveSnapshot(i *v1beta1.Instance) error {
+	jsonBytes, err := json.Marshal(i.Spec)
+	if err != nil {
+		return err
+	}
+	if i.Annotations == nil {
+		i.Annotations = make(map[string]string)
+	}
+	i.Annotations[snapshotAnnotation] = string(jsonBytes)
+	return nil
+}
+
+func snapshotSpec(i *v1beta1.Instance) (*v1beta1.InstanceSpec, error) {
+	if i.Annotations != nil {
+		snapshot, ok := i.Annotations[snapshotAnnotation]
+		if ok {
+			var spec *v1beta1.InstanceSpec
+			err := json.Unmarshal([]byte(snapshot), &spec)
+			if err != nil {
+				return nil, err
+			}
+			return spec, nil
+		}
+	}
+	return nil, nil
+}
+
+func remove(values []string, s string) (result []string) {
+	for _, value := range values {
+		if value == s {
+			continue
+		}
+		result = append(result, value)
+	}
+	return
+}
+
+// tryAddFinalizer adds the cleanup finalizer to an instance if the finalizer
+// hasn't been added yet, the instance has a cleanup plan and the cleanup plan
+// didn't run yet. Returns true if the cleanup finalizer has been added.
+func tryAddFinalizer(i *v1beta1.Instance) bool {
+	if !funk.ContainsString(i.ObjectMeta.Finalizers, instanceCleanupFinalizerName) {
+		if planStatus := i.PlanStatus(v1beta1.CleanupPlanName); planStatus != nil {
+			// avoid adding a finalizer again if a reconciliation is requested
+			// after it has just been removed but the instance isn't deleted yet
+			if planStatus.Status == v1beta1.ExecutionNeverRun {
+				i.ObjectMeta.Finalizers = append(i.ObjectMeta.Finalizers, instanceCleanupFinalizerName)
+				return true
+			}
+		}
+	}
+
+	return false
+}
+
+// tryRemoveFinalizer removes the cleanup finalizer of an instance if it has
+// been added, the instance has a cleanup plan and the cleanup plan completed.
+// Returns true if the cleanup finalizer has been removed.
+func tryRemoveFinalizer(i *v1beta1.Instance) bool {
+	if funk.ContainsString(i.ObjectMeta.Finalizers, instanceCleanupFinalizerName) {
+		if planStatus := i.PlanStatus(v1beta1.CleanupPlanName); planStatus != nil {
+			if planStatus.Status.IsTerminal() {
+				i.ObjectMeta.Finalizers = remove(i.ObjectMeta.Finalizers, instanceCleanupFinalizerName)
+				return true
+			}
+		} else {
+			// We have a finalizer but no cleanup plan. This could be due to an updated instance.
+			// Let's remove the finalizer.
+			i.ObjectMeta.Finalizers = remove(i.ObjectMeta.Finalizers, instanceCleanupFinalizerName)
+			return true
+		}
+	}
+
+	return false
 }
