+++ pkg/controller/instance/instance_controller_test.go
@@ -326,3 +326,25 @@ func Test_makePipes(t *testing.T) {
 		})
 	}
 }
+
+func TestSpecParameterDifference(t *testing.T) {
+	var testParams = []struct {
+		name string
+		new  map[string]string
+		diff map[string]string
+	}{
+		{"update one value", map[string]string{"one": "11", "two": "2"}, map[string]string{"one": "11"}},
+		{"update multiple values", map[string]string{"one": "11", "two": "22"}, map[string]string{"one": "11", "two": "22"}},
+		{"add new value", map[string]string{"one": "1", "two": "2", "three": "3"}, map[string]string{"three": "3"}},
+		{"remove one value", map[string]string{"one": "1"}, map[string]string{"two": "2"}},
+		{"no difference", map[string]string{"one": "1", "two": "2"}, map[string]string{}},
+		{"empty new map", map[string]string{}, map[string]string{"one": "1", "two": "2"}},
+	}
+
+	var old = map[string]string{"one": "1", "two": "2"}
+
+	for _, test := range testParams {
+		diff := parameterDiff(old, test.new)
+		assert.Equal(t, test.diff, diff)
+	}
+}
